refactor: 영화 목록 HomeFragment로 이동
