--- app/src/main/java/woowacourse/movie/feature/movies/view/MoviesActivity.kt
@@ -1,57 +1,0 @@
-package woowacourse.movie.feature.movies.view
-
-import android.os.Bundle
-import androidx.activity.enableEdgeToEdge
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
-import androidx.recyclerview.widget.RecyclerView
-import woowacourse.movie.R
-import woowacourse.movie.domain.model.Screening
-import woowacourse.movie.domain.model.Screenings
-import woowacourse.movie.feature.TheatersDialogFragment
-import woowacourse.movie.feature.bookingdetail.view.BookingDetailActivity
-import woowacourse.movie.feature.model.MovieUiModel
-import woowacourse.movie.feature.movies.contract.MoviesContract
-import woowacourse.movie.feature.movies.presenter.MoviesPresenter
-import woowacourse.movie.feature.movies.view.adapter.Item
-import woowacourse.movie.feature.movies.view.adapter.MoviesAdapter
-
-class MoviesActivity :
-    AppCompatActivity(),
-    MoviesContract.View {
-    private val presenter: MoviesContract.Presenter by lazy { MoviesPresenter(this) }
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setupView()
-        presenter.prepareMovies()
-    }
-
-    override fun showMovies(movies: List<MovieUiModel>) {
-        val moviesAdapter = MoviesAdapter { movie -> presenter.selectMovieForBooking(movie) }
-        moviesAdapter.submitList(Item.from(movies))
-        findViewById<RecyclerView>(R.id.rv_movies).adapter = moviesAdapter
-    }
-
-    override fun showTheaters(screenings: Screenings) {
-        TheatersDialogFragment(screenings) { screening ->
-            navigateToBookingDetail(screening)
-        }.show(supportFragmentManager, TheatersDialogFragment.TAG)
-    }
-
-    override fun navigateToBookingDetail(screening: Screening) {
-        val intent = BookingDetailActivity.newIntent(this, screening)
-        startActivity(intent)
-    }
-
-    private fun setupView() {
-        enableEdgeToEdge()
-        setContentView(R.layout.activity_movies)
-        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.activity_movies)) { v, insets ->
-            val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
-            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
-            insets
-        }
-    }
-}
